PO-3 - finished ProductDto and all required staff for it.

+++ core/src/main/java/priceobserver/data/productproperties/ProductProperties.java
@@ -20,13 +20,13 @@ public class ProductProperties {
     private String properties;
 
     @Version
-    private Long version;
+    private Integer version;
 
-    public Long getVersion() {
+    public Integer getVersion() {
         return version;
     }
 
-    public void setVersion(Long version) {
+    public void setVersion(Integer version) {
         this.version = version;
     }
